requirements.txt
docopt==0.6.2
PyYAML==3.12
requests==2.19.1
